Fix Core.c building errors.

+++ Core/Graphics/Graphics.h
@@ -11,7 +11,7 @@ void Graphics_PPU_Reset();
 
 void Graphics_PPU_Power_Cycle();
 
-uint8_t* Graphics_Get_Frame();
+uint32_t* Graphics_Get_Frame();
 
 uint8_t Graphics_PPU_Register_Read(uint16_t Address);
